backend/requirements.txt
aiofile>=3.5.0
aiohttp>=3.7.4
async-dns>=1.1.8
chardet>=2.1.7
elasticsearch[async]>=7.10.0
eml-parser>=1.14.6
geoip2>=4.1.0
hpfeeds>=3.0.0
maxminddb>=2.0.3
motor>=3.0.0
netaddr>=0.8.0
pyfaup>=1.2
pymisp[email]>=2.4.162.1
pymongo>=4.2.0
pylibemu>=0.7.0
python-magic>=0.4.27
PyYAML>=5.4
sflock2
validators==0.18.2
varname==0.6.3
karton-core
